% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINTed-package.R
\docType{package}
\name{MINTed-package}
\alias{MINTed}
\alias{MINTed-package}
\title{MINTed: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
