% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINTed-package.R
\docType{package}
\name{MINTed-package}
\alias{MINTed}
\alias{MINTed-package}
\title{MINTed: Post-process MINTs results stored in a DuckDB database}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Cosmo Santoni \email{first.last@example.com}

}
\keyword{internal}
